refactor(errors): move classification to src/index for future telemetry support
